backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyJWT
python-dotenv
psycopg2-binary
pydantic
pydantic-settings
email-validator
requests
paho-mqtt
